+++ src/enums/serviceMap.ts
@@ -97,6 +97,8 @@ import SystemsManagerInstance from '../services/systemsManagerInstance'
 import SystemsManagerDocument from '../services/systemsManagerDocument'
 import RdsClusterSnapshot from '../services/rdsClusterSnapshot'
 import VpcEndpoint from '../services/vpcEndpoint'
+import APIGatewayDomainName from '../services/apiGatewayDomainName'
+import APIGatewayHttpApi from '../services/apiGatewayHttpApi'
 
 /**
  * serviceMap is an object that contains all currently supported services for AWS
@@ -106,6 +108,8 @@ export default {
   account: Account,
   [services.appSync]: AppSync,
   [services.alb]: ALB,
+  [services.apiGatewayDomainName]: APIGatewayDomainName,
+  [services.apiGatewayHttpApi]: APIGatewayHttpApi,
   [services.apiGatewayResource]: APIGatewayResource,
   [services.apiGatewayRestApi]: APIGatewayRestApi,
   [services.apiGatewayStage]: APIGatewayStage,