+++ src/services/apiGatewayDomainName/mutation.ts
@@ -1,0 +1,5 @@
+export default `mutation($input: [AddawsApiGatewayDomainNameInput!]!) {
+  addawsApiGatewayDomainName(input: $input, upsert: true) {
+    numUids
+  }
+}`