+++ src/services/apiGatewayHttpApi/mutation.ts
@@ -1,0 +1,5 @@
+export default `mutation($input: [AddawsApiGatewayHttpApiInput!]!) {
+  addawsApiGatewayHttpApi(input: $input, upsert: true) {
+    numUids
+  }
+}`